Move PhonesConfig helpers to phones.app
